CI Github workflow for Arduino and PlatformIO
Build the project using the current PlatformIO and Arduino
development environments on every push and pull request.

The Arduino build installs all of the dependencies manually before
building the sketch.

The PlatformIO build installs all dependencies automatically before
building the debug and release builds of the project.

The README is updated to add a status badge to allow the continuous
integration status to be seen easily.