Fix structure initialisation in ESP-IDF structures
The uart_config_t structure used in the main sketch setup() was
not fully initialised. The fix uses the correct constants from
the ESP-IDF to represent the zero value the original code was
implicitly assuming.

The twai_general_config_t structure in OICan::Init() was missing
an interrupt flag initialiser. Fix the warning by explicitly
setting to zero to match the previous implict assumption.